backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic-settings==2.3.4
anthropic==0.34.2
pydicom==2.4.4
SimpleITK==2.3.1
nibabel==5.2.1
numpy==1.26.4
scipy==1.13.1
pyvista==0.43.10
vtk==9.3.1
trimesh==4.4.9
pillow==10.4.0
pymupdf==1.24.10
aiofiles==23.2.1
python-dotenv==1.0.1
